Gatsby fixes and annoatated docs
