+++ packages/components-date/README.md
@@ -1,5 +1,3 @@
 # Looker UI Date Components
 
-## Introduction
-
 This package is a supplement to @looker/components with some Date-related components that depend on `date-fns` and `react-date-picker`
