+++ www/src/MDX/Pre/allComponents.ts
@@ -168,7 +168,6 @@ import {
   VisuallyHidden,
   i18nResources,
 } from '@looker/components'
-
 import {
   Calendar,
   DateFormat,
@@ -182,7 +181,7 @@ import {
   InputTime,
   InputTimeSelect,
   TimeFormat,
-} from '@looker/components/src/date'
+} from '@looker/components-date'
 
 import { DialogSurface } from '@looker/components/src/Dialog/DialogSurface'
