+++ www/src/documentation/components/date/calendar.mdx
@@ -4,6 +4,9 @@ title: Calendar
 
 import { Code } from '@looker/components'
 import { CalendarPropTable } from './demos'
+import { Message } from './ComponentsDate'
+
+<Message>Calendar</Message>
 
 `Calendar` is a stateless component which renders a standard calendar month. It accepts a limited amount of props, and is the foundation of our more useful [`InputDate`](/components/forms/input-date) and `InputDateRange` components.
