+++ www/src/documentation/components/date/date-format.mdx
@@ -4,6 +4,9 @@ github: DateTimeFormat/DateFormat.tsx
 ---
 
 import { DateFormatTable, LocaleTable } from './demos'
+import { Message } from './ComponentsDate'
+
+<Message>DateFormat</Message>
 
 As it's name implies, `DateFormat` is used to format a javascript Date object to human readable and localized text.
