+++ www/src/documentation/components/date/demos/LocaleTable.tsx
