+++ www/src/documentation/components/date/demos/index.ts
@@ -24,8 +24,7 @@
 
  */
 
-export * from './IconList'
-export * from './SelectManagerParameterTable'
-export * from './SelectManagerReturnTable'
-export * from './NamedBreakpointTable'
-export * from './ListItemPropTable'
+export * from './CalendarPropTable'
+export * from './DateFormatTable'
+export * from './LocaleTable'
+export * from './TimeFormatTable'
