+++ www/src/documentation/components/date/field-date-range.mdx
@@ -4,7 +4,9 @@ github: FieldDateRange
 storybook: true
 ---
 
-Import path: `@looker/components/lib/FieldDateRange`
+import { Message } from './ComponentsDate'
+
+<Message>FieldDateRange</Message>
 
 This component is composed of an `InputDateRange` and a `Label`.
