+++ www/src/documentation/components/date/field-date.mdx
@@ -3,7 +3,9 @@ title: FieldDate
 storybook: true
 ---
 
-Import path: `@looker/components/lib/FieldDate`
+import { Message } from './ComponentsDate'
+
+<Message>FieldDate</Message>
 
 This component is composed of an `InputDate` and a `Label`.
