+++ www/src/documentation/components/date/field-time.mdx
@@ -4,6 +4,10 @@ github: Form/Fields/FieldTime
 storybook: true
 ---
 
+import { Message } from './ComponentsDate'
+
+<Message>FieldTime</Message>
+
 The `<FieldTime />` component is composed of an `<InputTime />` component and a `<Label />` component. It has attributes to help display additional information. `detail`, `description` and `validation messages` are available to the component.
 
 ```jsx
