+++ www/src/documentation/components/date/input-date-range.mdx
@@ -4,6 +4,10 @@ github: Form/Inputs/InputDateRange/InputDateRange.tsx
 storybook: true
 ---
 
+import { Message } from './ComponentsDate'
+
+<Message>InputDateRange</Message>
+
 `InputDateRange` provides an interface to select to a date range, either through the calendar interface or text input boxes.
 
 ## Value
