+++ www/src/documentation/components/date/input-date.mdx
@@ -4,6 +4,10 @@ github: Form/Inputs/InputDate/InputDate.tsx
 storybook: true
 ---
 
+import { Message } from './ComponentsDate'
+
+<Message>InputDate</Message>
+
 `InputDate` provides a text input and calendar to select a single date, which can be used as a form input or filter.
 
 ## Value & Default Value
