+++ www/src/documentation/components/date/input-time-select.mdx
@@ -3,6 +3,10 @@ title: InputTimeSelect
 github: Form/Inputs/InputTimeSelect/InputTimeSelect.tsx
 ---
 
+import { Message } from './ComponentsDate'
+
+<Message>InputTimeSelect</Message>
+
 `InputTimeSelect` provides a [`Select`](/components/forms/select)-based dropdown UI for picking a 24-hour time value. This is an alternative to [`InputTime`](/components/forms/input-time), which provides an HTML5-standard time picking experience.
 
 The default implementation provides a dropdown list of times in 15-minute increments, formatted in for a 12-hour clock.
