+++ www/src/documentation/components/date/input-time.mdx
@@ -3,6 +3,10 @@ title: InputTime
 github: Form/Inputs/InputTime/InputTime.tsx
 ---
 
+import { Message } from './ComponentsDate'
+
+<Message>InputTime</Message>
+
 `InputTime` provides a themable HTML5-equivalent time picking experience.
 
 ```js
