+++ www/src/documentation/components/date/time-format.mdx
@@ -4,6 +4,9 @@ github: DateTimeFormat/TimeFormat.tsx
 ---
 
 import { TimeFormatTable, LocaleTable } from './demos'
+import { Message } from './ComponentsDate'
+
+<Message>TimeFormat</Message>
 
 As it's name implies, `TimeFormat` is used to format a javascript Date object to human readable and localized text.
